requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
anytree==2.9.0
    # via -r requirements.in
asttokens==2.4.0
    # via stack-data
backcall==0.2.0
    # via ipython
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
exceptiongroup==1.1.3
    # via ipython
executing==1.2.0
    # via stack-data
ipdb==0.13.13
    # via -r requirements.in
ipython==8.15.0
    # via ipdb
jedi==0.19.0
    # via ipython
lxml==4.9.3
    # via metsrw
matplotlib-inline==0.1.6
    # via ipython
metsrw==0.4.0
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.39
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.16.1
    # via ipython
pymssql==2.2.8
    # via -r requirements.in
python-slugify==8.0.1
    # via -r requirements.in
rapidfuzz==3.5.2
    # via thefuzz
six==1.16.0
    # via
    #   anytree
    #   asttokens
stack-data==0.6.2
    # via ipython
text-unidecode==1.3
    # via python-slugify
thefuzz==0.20.0
    # via -r requirements.in
tomli==2.0.1
    # via ipdb
traitlets==5.10.0
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.6
    # via prompt-toolkit
